+++ my-project/Day26/src/app/boards/[boardId]/edit/page.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import BoardsNew from "@/components/boards-write";
 
 export default function EditPage() {
