+++ my-project/Day26/src/components/boards-detail/detail/queries.ts
@@ -1,0 +1,15 @@
+import { gql } from "@apollo/client";
+
+const FETCH_BOARD = gql`
+  query fetchBoard($boardId: ID!) {
+    fetchBoard(boardId: $boardId) {
+      _id
+      writer
+      title
+      contents
+      createdAt
+    }
+  }
+`;
+
+export { FETCH_BOARD };
