+++ my-project/Day26/src/components/boards-list/queries.ts
@@ -1,0 +1,20 @@
+import { gql } from "@apollo/client";
+
+const FETCH_BOARDS = gql`
+  query {
+    fetchBoards(page: 1) {
+      _id
+      writer
+      title
+      createdAt
+    }
+  }
+`;
+
+const DELETE_BOARD = gql`
+  mutation deleteBoard($boardId: ID!) {
+    deleteBoard(boardId: $boardId)
+  }
+`;
+
+export { FETCH_BOARDS, DELETE_BOARD };
