+++ my-project/Day26/src/components/boards-write/queries.ts
@@ -1,0 +1,56 @@
+import { gql } from "@apollo/client";
+
+const CREATE_BOARD = gql`
+  mutation createBoard($createBoardInput: CreateBoardInput!) {
+    createBoard(createBoardInput: $createBoardInput) {
+      _id
+      writer
+      title
+      contents
+    }
+  }
+`;
+
+const FETCH_BOARD = gql`
+  query fetchBoard($boardId: ID!) {
+    fetchBoard(boardId: $boardId) {
+      _id
+      writer
+      title
+      contents
+      createdAt
+    }
+  }
+`;
+
+const FETCH_BOARDS = gql`
+  mutation fetchBoards($page: int) {
+    fetchBoards(page: $page) {
+      _id
+      writer
+      title
+      createdAt
+    }
+  }
+`;
+
+const UPDATE_BOARD = gql`
+  mutation updateBoard(
+    $boardId: ID!
+    $password: String
+    $updateBoardInput: UpdateBoardInput!
+  ) {
+    updateBoard(
+      boardId: $boardId
+      password: $password
+      updateBoardInput: $updateBoardInput
+    ) {
+      _id
+      writer
+      title
+      contents
+    }
+  }
+`;
+
+export { CREATE_BOARD, FETCH_BOARD, FETCH_BOARDS, UPDATE_BOARD };
